Refactor(step4) 우승자를 확인하는 역할을 Referee 객체로 이동, ResultView에서 해당 역할 제거
