+++ src/main/kotlin/step4/model/Car.kt
@@ -1,7 +1,7 @@
 package step4.model
 
 data class Car(val name: String, var raceResult: Int, val raceHistory: MutableList<Int>) {
-    fun saveResult(resultOfTurn: Int) {
+    fun save(resultOfTurn: Int) {
         raceResult += resultOfTurn
         raceHistory.add(resultOfTurn)
     }
